Add user information to Alert model and create UserAlertsListDeleteView
The commit includes the addition of a 'user' field to the Alert model in 'models.py'. Moreover, the AlertSerializer and AlertSerializerSave in 'serializers.py' now include 'user' in their fields. A new view for listing and deleting user-specific alerts, UserAlertsListDeleteView, is added to 'views.py'. This change also gets reflected in 'urls.py'. These updates primarily enhance the Alert management in the application to be user-specific.